share data across kernels.