@ReinForce-II 2020-11-22
benchmark: super_resolution_10

5900X:
before:
Constant-10      default         10        0.000(us)
Conv-10          default         20  1496819.700(us)
Relu-10          default         15      964.000(us)
Reshape-10       default         10      124.800(us)
Transpose-10     default         5     6619.800(us)

after:
Constant-10      default         10        0.200(us)
Conv-10          default         20   885816.550(us)
Relu-10          default         15      847.733(us)
Reshape-10       default         10      129.800(us)
Transpose-10     default         5     6731.200(us)

M1:
before:
Constant-10      default         10        0.000(us)
Conv-10          default         20   650797.950(us)
Relu-10          default         15      411.600(us)
Reshape-10       default         10       37.600(us)
Transpose-10     default         5     3822.400(us)

after:
Constant-10      default         10        0.100(us)
Conv-10          default         20   651256.650(us)
Relu-10          default         15      446.067(us)
Reshape-10       default         10       40.200(us)
Transpose-10     default         5     3787.400(us)

RK3399:
A72:
before:
Constant-10      default         2        0.500(us)
Conv-10          default         4  4831419.250(us)
Relu-10          default         3     2834.000(us)
Reshape-10       default         2      504.000(us)
Transpose-10     default         1    18659.000(us)

after:
Constant-10      default         2        0.000(us)
Conv-10          default         4  3472175.500(us)
Relu-10          default         3     3046.333(us)
Reshape-10       default         2      613.500(us)
Transpose-10     default         1    18683.000(us)

A53:
before:
Constant-10      default         2        1.000(us)
Conv-10          default         4 15421956.500(us)
Relu-10          default         3     6136.667(us)
Reshape-10       default         2      971.000(us)
Transpose-10     default         1    47123.000(us)

after:
Constant-10      default         2        0.500(us)
Conv-10          default         4  9812648.000(us)
Relu-10          default         3     6146.000(us)
Reshape-10       default         2     1008.500(us)
Transpose-10     default         1    47235.000(us)